+++ clang-tools-extra/clang-tidy/hooks/HooksTidyModule.cpp
@@ -31,6 +31,7 @@
 #include "FloatManipPureCheck.h"
 #include "FloatOnePureCheck.h"
 #include "FloatPureCheck.h"
+#include "FuncAddrTakenCheck.h"
 #include "GuardCalledCheck.h"
 #include "GuardInForCheck.h"
 #include "GuardInWhileCheck.h"
@@ -101,6 +102,8 @@ public:
         "hooks-float-one-pure");
     CheckFactories.registerCheck<FloatPureCheck>(
         "hooks-float-pure");
+    CheckFactories.registerCheck<FuncAddrTakenCheck>(
+        "hooks-func-addr-taken");
     CheckFactories.registerCheck<GuardCalledCheck>(
 	"hooks-guard-called");
     CheckFactories.registerCheck<GuardInForCheck>(